move some logical size bits to separate logical_size.rs
